device_control_app/requirements.txt
anyio==3.6.1
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
fastapi==0.78.0
h11==0.13.0
idna==3.3
importlib-metadata==4.11.4
pkg-resources==0.0.0
PyBluez==0.23
pydantic==1.9.1
python-dotenv==0.20.0
requests==2.28.0
sniffio==1.2.0
starlette==0.19.1
typing-extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
zipp==3.8.0
